Move rotate to a package.
