+++ dolfyn/rotate/awac.py
@@ -1,2 +1,2 @@
 from .vector import inst2earth, earth2principal
-from dolfyn.rdi.rotate import beam2inst
+from rdi import beam2inst
