+++ dolfyn/rotate/base_nortek.py
